Refresh property list (language bar) only when needed
When users set the property list (language bar) to "auto hide"
in ibus-setup, the bar will stick to the cursor and block the
candidats list, because ibus-chewing refreshes the bar on every
keystroke. Therefore, it's better to refresh the bar only when
users toggle Chi-Eng Mode or Shape Mode with Shift or Caps Lock.