Fix: Ensure Ctrl+Space Is Passed to IBus Correctly, Not Absorbed
On GNOME 46 (and maybe earlier versions) with Wayland, the Ctrl+Space key event somehow cannot be passed to IBus correctly, which prevents users from switching input methods with that shortcut. 

As this event does not need to be handled by libchewing, we simply remove Ctrl from the modifiers for the self_handle_space function. This will ensure the event is always passed to IBus.